backend/requirements.txt
fastapi 
uvicorn 
motor 
python-dotenv 
passlib[bcrypt]  
PyJWT
pydantic[email]
llama_index
llama_index-core
llama_index-embeddings-gemini
llama_index-llms-gemini
Pypdf2
pyvis
python-multipart
motor
pymongo
google-generativeai
PyMuPDF
python-jose
google-genai
matplotlib
aiohttp
python-docx
docx